+++ field/src/extension/mod.rs
@@ -1,5 +1,4 @@
 use crate::field::Field;
-use crate::AbstractExtensionField;
 
 pub mod binomial_extension;
 pub mod cubic;
@@ -17,41 +16,4 @@ pub trait BinomiallyExtendable<const D: usize>: Field + Sized {
     const DTH_ROOT: Self;
 
     fn ext_multiplicative_group_generator() -> [Self; D];
-}
-
-pub trait Frobenius<F: BinomiallyExtendable<D>, const D: usize>:
-    Field + Sized + AbstractExtensionField<F>
-{
-    /// FrobeniusField automorphisms: x -> x^n, where n is the order of BaseField.
-    fn frobenius(&self) -> Self {
-        self.repeated_frobenius(1)
-    }
-
-    /// Repeated Frobenius automorphisms: x -> x^(n^count).
-    ///
-    /// Follows precomputation suggestion in Section 11.3.3 of the
-    /// Handbook of Elliptic and Hyperelliptic Curve Cryptography.
-    fn repeated_frobenius(&self, count: usize) -> Self {
-        if count == 0 {
-            return *self;
-        } else if count >= D {
-            // x |-> x^(n^D) is the identity, so x^(n^count) ==
-            // x^(n^(count % D))
-            return self.repeated_frobenius(count % D);
-        }
-        let arr: &[F] = self.as_base_slice();
-
-        // z0 = DTH_ROOT^count = W^(k * count) where k = floor((n-1)/D)
-        let mut z0 = F::DTH_ROOT;
-        for _ in 1..count {
-            z0 *= F::DTH_ROOT;
-        }
-
-        let mut res = [F::ZERO; D];
-        for (i, z) in z0.powers().take(D).enumerate() {
-            res[i] = arr[i] * z;
-        }
-
-        Self::from_base_slice(&res)
-    }
 }
